+++ scripts/main.js
@@ -10,6 +10,7 @@ import "../scripts/witnesses/witnessDataProvider.js"
 import { getWitnesses } from "../scripts/witnesses/witnessDataProvider.js";
 // import { getCriminals } from "./criminals/CriminalProvider.js";
 import { ShowWitnessButton } from "../scripts/witnesses/witnessButton.js";
+import "./witnesses/witnessList.js"
 
 OfficerList()
 CriminalList()
