+++ scripts/witnesses/witness.js
@@ -1,9 +1,8 @@
-// export const Witness = (witness) => {
-//     return `
-//     <p class="witness"> 
-//         <b>${witness.name}</b> <br>
-//         <b>ID: </b>${witness.id} <br>
-//         <b>Statements: </b>${witness.statements}<br>
-//     </p>    
-//     `
-// }
+export const Witness = (witness) => {
+    return `
+    <p class="witness"> 
+        <b>Witness Name: </b>${witness.name}<br>
+        <b>Statements: </b>${witness.statements}<br>
+    </p>    
+    `
+}
