+++ scripts/witnesses/witnessDataProvider.js
