+++ scripts/witnesses/witnessList.js
@@ -1,8 +1,32 @@
-eventHub.addEventListener()
+import { Witness } from "./Witness.js"
+import { getWitnesses, useWitness } from "./WitnessDataProvider.js"
+import "./witnessButton.js"
 
-// eventHub.addEventListener("click", clickEvent => {
-//     if (clickEvent.target.id === "showWitnesses") {
-//         const customEvent = new CustomEvent("showWitnessStatements")
-//         eventHub.dispatchEvent(customEvent)
-//     }
-// })
+const eventHub = document.querySelector(".container")
+const contentContainer = document.querySelector(".witnessContainer")
+const criminalContainer = document.querySelector(".criminalsContainer")
+
+eventHub.addEventListener("showWitnessesClicked", event => {
+    WitnessList()
+})
+
+export const WitnessList = () => {
+    criminalContainer.innerHTML = ""
+    getWitnesses()
+    .then(() => {
+        const witnessesArray = useWitness()
+        debugger
+        renderToDom(witnessesArray)
+    })
+}
+
+const renderToDom = (witnessCollection) => {
+    let witnessHTMLrepresentation = ""
+    for (const witnessObj of witnessCollection) {
+            witnessHTMLrepresentation += Witness(witnessObj)
+        }
+        contentContainer.innerHTML = `
+            <h2>Witness Statements</h2>
+            ${witnessHTMLrepresentation}
+            </section>`
+        }
